backend/requirements.txt

fastapi
uvicorn
pytesseract
Pillow
requests
python-multipart
sqlite-utils
